backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
sqlalchemy==2.0.30
psycopg[binary]==3.1.19
pydantic==1.10.15
pyjwt==2.8.0
celery==5.4.0
redis==5.0.4
pytest==8.2.1
